Moved ApplicationChoiceStatusTagComponent
